Move gRPC type convertion functions to separate file
